Renamed IntegerMatrix class to Statistics. Removed variable size constructor and implemented getStat(index) method to pull base+bonus of a stat from the matrix.
Created BattleEffects class and wrote the basic constructor.
Moved player.level to entity.level so enemies can have levels. Added checkLvlUp method in player to be called at the end of battles.
Added another constructor to be used in entity.do(), also added getEntityStats in Ability class so that damage can be calculated with the entity's stats.